Fixed not linking files.
This fixes where files don't get linked unless they match.  Files that
match are used for buiding XML to HTML.  If linking is used, all files
that qualify (are included and are not excluded) should be linked, even
if they don't match for building.  This way, resources will be linked in
the output directories.